Use a single SimpleProvider instead of a provider implementation per client.
With the currently implemented clients, all that is needed is a builder with the client configuration to successfully build a client. This means that a provider implementation that accepts a generic builder trait is enough to build any of the existing clients.

In the future, other providers may be added depending on data that can be shared between multiple instances of active clients.